pyproject.toml
[project]
name = "rer-api-wrapper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "playwright>=1.58.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "google-auth-oauthlib>=1.2.0",
    "google-auth>=2.27.0",
    "google-api-python-client>=2.111.0",
    "beautifulsoup4>=4.14.3",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
